TODO cleanup, file name cleanup, monday night PR push
