+++ client/src/App.js
@@ -9,8 +9,7 @@ import {
   Switch
 } from "react-router-dom";
 
-import NavBar from "./components/NavBar/navbar.js";
-import Event from "./components/Event";
+import NavBar from "./components/NavBar";
 import 'semantic-ui-css/semantic.min.css';
 
 import LandingPage from "./pages/LandingPage";
@@ -19,7 +18,7 @@ import ProfilePage from "./pages/ProfilePage";
 import CreateEventPage from "./pages/CreateEventPage";
 import EventProfilePage from "./pages/EventProfilePage";
 import NotFoundPage from "./pages/NotFoundPage";
-import UnsponsoredEventsPage from "./pages/UnsponsoredEventsPage"
+import UnsponsoredEventsPage from "./pages/UnsponsoredEventsPage";
 
 import AuthLogin from "./components/Auth/Login";
 
@@ -41,7 +40,7 @@ function App() {
           <Route exact path="/login" component={LandingPage} />
           <Route exact path="/register" component={RegisterPage} />
           <Route exact path="/unsponsored" component={UnsponsoredEventsPage} />
-          <PrivateRoute exact path="/event" component={Event} />
+          <PrivateRoute exact path="/event" component={EventProfilePage} />
           <PrivateRoute path="/profile/:id" component={ProfilePage} />
           <PrivateRoute exact path="/createevent" component={CreateEventPage} />
           <Route path="/event/:id" component={EventProfilePage} />
