--- client/src/components/Event/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Event() {
-    return (<h1>Event Component</h1>)
-}
-
-export default Event;
